fix: gcc and clang compile warnings (#15)
# Main Changes
1. modified bazel to expose the warnings when compiling the unit tests.
2. fixed the compiling warnings exposed in step 1.

+++ include/sonic/internal/skip.h
@@ -450,7 +450,8 @@ class SkipScanner {
       goto err_invalid_char;
     }
     // match key and skip parsing unneeded fields
-    if (sn == key.size() && std::memcmp(sp, key.data(), sn) == 0) {
+    if (sn == static_cast<long>(key.size()) &&
+        std::memcmp(sp, key.data(), sn) == 0) {
       goto query;
     } else {
       c = SkipSpaceSafe(data, pos, len);
